+++ Editor/Flow/Templates/FlowNode_ForEachLoopT_Template.cs
@@ -1,0 +1,7 @@
+﻿namespace Ceres.Editor.Graph.Flow
+{
+    internal sealed class FlowNode_ForEachLoopT_Template: IReadOnlyListNodeTemplate
+    {
+        
+    }
+}
